// apps/frontend/src/client/types.gen.ts
export type ConnectionConfig = {
	/**
	 * The database management system hosting the target database. Defaults to PostgreSQL.
	 */
	dbms?: Dbms;
	/**
	 * Username
	 *
	 * The username used to authenticate with the database server.
	 */
	username: string;
	/**
	 * Password
	 *
	 * The password used to authenticate with the database server.
	 */
	password: string;
	/**
	 * Host
	 *
	 * The hostname or IP address of the database server.
	 */
	host: string;
	/**
	 * Port
	 *
	 * The port number the database server is listening on.
	 */
	port: number;
	/**
	 * Db
	 *
	 * The name of the database to connect to.
	 */
	db: string;
	/**
	 * Schema
	 *
	 * The schema to set as the search path. Only applicable to database systems that support schemas,such as PostgreSQL. Leave empty for others.
	 */
	schema?: string | null;
};

/**
 * ConnectionStatus
 */
export type ConnectionStatus = {
	/**
	 * Connected
	 */
	connected: boolean;
	/**
	 * Database
	 *
	 * Name of the connected database, or 'null' if not connected.
	 */
	database?: string | null;
};

/**
 * DBMS
 *
 * Supported database management systems and their SQLAlchemy URL schemes.
 */
export type Dbms =
	| "postgresql"
	| "mysql"
	| "mariadb"
	| "sqlite"
	| "oracle"
	| "mssql";